+++ src/main/java/spacewars/view/menus/how_to_play/SpacebarViewer.java
@@ -13,7 +13,7 @@ public class SpacebarViewer {
         this.image = imageLoader.get("images/spacebar_howtoplay.png");
     }
 
-    public void draw(GUI gui, int x, int y) throws IOException {
+    public void draw(GUI gui, double x, double y) throws IOException {
         image.draw(gui, x, y);
     }
 }
